Refactor: filter 기능 분리
